Retinoblastoma detector with a picture or live capture
The Retinoblastoma_Detector(face).py takes in a picture and puts a red box around the eye with Retinoblastoma and green around normal eyes. The Retinoblastoma_Detector(video).py does about the same thing but instead of using an image, it uses the webcam and it keeps scanning. At the moment both of these programs are very inaccurate because its cropping too big of an image of the eye, throwing off the prediction. I am working on this currently so it should be fixed soon. Also with the live capture program using the esc key to end the program doesn't work for some reason and I am working on fixing that.